Enable PoolBlockingPeriod for .NET Standard + expose GetColumnSchema
